requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
email_validator==2.2.0
fastapi==0.115.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
instructor==1.5.0
jiter==0.5.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
openai==1.50.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
rich==13.8.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.38.6
tenacity==8.5.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.31.0
yarl==1.13.1
